+++ src/RefreshAheadCacheBehavior.php
@@ -109,11 +109,11 @@ class RefreshAheadCacheBehavior extends Behavior
     private $_refreshTimeoutCache;
 
     /**
-     * Resolves the specified reference into an actual [[RefreshAheadConfig]]
+     * Resolves the specified reference into an actual [[CallableGenerator]]
      * object.
      *
      * The reference may be specified as a configuration array for a
-     * [[RefreshAheadConfig]] object. In this case, you must specify both
+     * [[CallableGenerator]] object. In this case, you must specify both
      * `refresh` and `generate` callables.
      *
      * The reference may be specified as a callable or \Closure, which will be
@@ -123,16 +123,16 @@ class RefreshAheadCacheBehavior extends Behavior
      * The reference may also be a string or an Instance object. If the former,
      * it will be treated as an application component ID or a class name.
      *
-     * @param  callable|\Closure|array|RefreshAheadConfig|string|Instance $reference an
+     * @param  callable|\Closure|array|CallableGenerator|string|Instance $reference an
      *     object or reference to the desired object.
      *
      *     You may specify a reference in terms of an application component ID,
-     *     an Instance object, a RefreshAheadConfig object, or a configuration
+     *     an Instance object, a CallableGenerator object, or a configuration
      *     array for creating the object. If the "class" value is not specified
      *     in the configuration array, it will use the value of
-     *     'thamtech\caching\refreshahead\RefreshAheadConfig'.
+     *     'thamtech\caching\refreshAhead\CallableGenerator'.
      *
-     * @return RefreshAheadConfig the object instance
+     * @return CallableGenerator the object instance
      * @throws InvalidConfigException if the reference is invalid
      */
     public static function ensureGenerator($reference)
@@ -143,7 +143,7 @@ class RefreshAheadCacheBehavior extends Behavior
             ];
         }
 
-        return Instance::ensure($reference, RefreshAheadConfig::class);
+        return Instance::ensure($reference, CallableGenerator::class);
     }
 
     /**