+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 ### Changed
 - Refactor: Extract Interface `GeneratorInterface` from `RefreshAheadConfig`
 - Move function `RefreshAheadConfig::ensure()` to `RefreshAheadCacheBehavior::ensureGenerator()`
+- Rename `RefreshAheadConfig` to `CallableGenerator`
 
 
 [v0.2.0]